pyproject.toml
[project]
name = "diffusion-trajectory-forecaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyOpenSSL>=24.0.0",
    "cryptography>=42.0.0",
    "dvc-webdav>=3.0.1",
    "dvc[gdrive,s3]>=3.67.0",
    "einops>=0.8.2",
    "equinox>=0.13.4",
    "google-auth>=2.48.0",
    "hydra-core>=1.3.2",
    "jax[cuda12]>=0.9.0.1",
    "jaxlib>=0.9.0.1",
    "lightning>=2.6.1",
    "matplotlib>=3.10.8",
    "optuna>=4.4.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "torch>=2.10.0",
    "webdataset>=0.2.111",
    "nvidia-cudnn-cu12==9.20.0.48",
    "clearml>=1.16.5",
    "waymo-waymax",
    "tensorboard>=2.20.0",
    "setuptools<81",
    "tensorboard-plugin-profile<2.20",
    "pandas>=3.0.2",
]

[tool.uv.sources]
waymo-waymax = { git = "https://github.com/waymo-research/waymax.git", rev = "main" }
